samples/python/requirements.txt
flask==0.12.0
django==1.11.0
requests==2.20.0
pyyaml==5.1.0
sqlalchemy==1.3.0
jinja2==2.10.0
